chore: move related layout componets to components/layout/ dir
