+++ frontend/src/components/Header.tsx
@@ -1,4 +1,11 @@
-import type { HeaderProps } from 'additional'
+import type { AriaAttributes } from "react"
+import type { DropUndefined } from "utilities"
+
+export interface HeaderProps {
+  children: React.ReactNode
+  labelledlby: DropUndefined<AriaAttributes, 'aria-labelledby'>
+  className: string
+}
 
 export const Header: React.FC<HeaderProps> = ({ children, labelledlby, className }) => {
   return (
