+++ frontend/src/components/layout/refs/index.ts
@@ -1,0 +1,2 @@
+export * from './headerLinks.ref'
+export * from './footerLinks.ref'
