+++ frontend/src/refs.ts
@@ -1,4 +1,3 @@
-import type { FooterLinks, HeaderLinks } from 'additional'
 import { ProductColors } from 'enums'
 
 export const productColor: Record<ProductColors, string> = Object.freeze({
@@ -23,93 +22,3 @@ export const filterStructure = {
   brand: null,
   discount: null
 }
-
-export const headerLinks: HeaderLinks[] = [
-  {
-    path: '/',
-    name: 'Home'
-  },
-  {
-    path: '/about',
-    name: 'About'
-  },
-  {
-    path: '#categories',
-    name: 'Categories',
-    dropdown: [
-      {
-        path: "/men's clothing",
-        name: "Men's clothing"
-      },
-      {
-        path: "/women's clothing",
-        name: "Women's clothing"
-      },
-      {
-        path: '/shoes',
-        name: 'Shoes'
-      },
-      {
-        path: '/accessory',
-        name: 'Accesories'
-      }
-    ]
-  }
-]
-
-export const footerLinks: FooterLinks[] = [
-  {
-    name: 'About',
-    path: '',
-    links: [
-      {
-        name: 'Features',
-        path: '#'
-      },
-      {
-        name: 'Pricing',
-        path: '#'
-      },
-      {
-        name: 'Payments methods',
-        path: '#'
-      }
-    ]
-  },
-  {
-    name: 'Policy',
-    path: '',
-    links: [
-      {
-        name: 'Privacy Policy',
-        path: '#'
-      },
-      {
-        name: 'Terms of Service',
-        path: '#'
-      },
-      {
-        name: 'Consumer Rights',
-        path: '#'
-      }
-    ]
-  },
-  {
-    name: 'Help',
-    path: '',
-    links: [
-      {
-        name: 'Common Questions',
-        path: '#'
-      },
-      {
-        name: 'Support',
-        path: '#'
-      },
-      {
-        name: 'Contact us',
-        path: '#'
-      }
-    ]
-  }
-]
